Refactor: parse_dir - No longer causes panic
parse_dir caused a panic as it did not check to see if the next path
being passed to `parse()` was a file or a directory. When `parse()` is
given a directory it will panic due the macro. This will need to be
refactored out to return an IO Error perhaps.