Refactor: parse_dir and parse
- Fix: parse_dir() this now does not try to call parse() on directory
  below itself.
- Fix: parse() now returns an error if a directory is parsed to it
  directly.

Note: Because parse_dir uses `path.is_dir()` `read_dir()` does not throw
any errors. It might be better in the future to remove this guard and
return error correctly so that crate uses to can be aware of IO errors.
If I am uderstand the results I am seeing. As the `unreadable_dir` test
does not return any errors in the current test cases.